Enforce clang-tidy (#373)

+++ include/core/hist/detail/data/WidthControllers.h
@@ -12,7 +12,7 @@ namespace ausaxs::hist::detail {
     };
 
     struct VariableWidth {
-        static float get() {return settings::flags::inv_bin_width;}
+        static float get() {return static_cast<float>(settings::flags::inv_bin_width);}
     };
 
     template<bool variable_bin_width>
