+++ include/math/math/MovingAverager.h
@@ -3,9 +3,9 @@
 
 #pragma once
 
-#include <vector>
-#include <cmath>
 #include <cassert>
+#include <cmath>
+#include <vector>
 
 namespace ausaxs {
     struct MovingAverage {
@@ -19,7 +19,7 @@ namespace ausaxs {
              * @param window_size The window size. 
              */
             template<typename T>
-            static std::vector<double> average(const T& data, unsigned int window_size) {
+            static std::vector<double> average(const T& data, int window_size) {
                 validate_input(data.size(), window_size);
                 return weighted_average(data, std::vector<double>(window_size, 1));
             }
@@ -34,14 +34,14 @@ namespace ausaxs {
              * @param window_size The window size. 
              */
             template<typename T>
-            static std::vector<double> average_half(const T& data, unsigned int window_size) {
+            static std::vector<double> average_half(const T& data, int window_size) {
                 validate_input(data.size(), window_size);
                 std::vector<double> weights(window_size, 1);
 
                 // define weights
-                unsigned int steps = (window_size-1) / 2;
-                unsigned int mid = steps;
-                for (unsigned int i = 1; i < steps+1; i++) {
+                int steps = (window_size-1) / 2;
+                int mid = steps;
+                for (int i = 1; i < steps+1; i++) {
                     double val = 1/std::pow(1.5, i);
                     weights[mid-i] = val;
                     weights[mid+i] = val;
@@ -51,15 +51,15 @@ namespace ausaxs {
             }
 
         private: 
-            static void validate_input(unsigned int N, unsigned int window_size);
+            static void validate_input(int N, int window_size);
 
             template<typename T>
             static std::vector<double> weighted_average(const T& data, std::vector<double> weights) {
                 std::size_t window_size = weights.size();
                 std::vector<double> averages(data.size());
-                unsigned int mid = (weights.size()-1)/2;
+                int mid = static_cast<int>((weights.size()-1)/2);
 
-                auto average = [&data, &weights, &mid] (unsigned int index, int steps) {
+                auto average = [&data, &weights, &mid] (int index, int steps) {
                     double sum = 0;
                     double w_sum = 0;
                     for (int j = -steps; j < steps+1; j++) {
@@ -70,12 +70,12 @@ namespace ausaxs {
                     return sum/w_sum;
                 };
 
-                unsigned int steps = (window_size-1)/2;
-                for (unsigned int i = steps; i < data.size() - steps; i++) {
+                int steps = static_cast<int>((window_size-1)/2);
+                for (int i = steps; i < static_cast<int>(data.size()) - steps; i++) {
                     averages[i] = average(i, steps);
                 }
 
-                for (unsigned int i = 0; i < steps; i++) {
+                for (int i = 0; i < steps; i++) {
                     averages[i] = average(i, i);
                     averages[data.size()-1-i] = average(data.size()-1-i, i);
                 }
