+++ source/core/mini/detail/Landscape.cpp
@@ -2,6 +2,7 @@
 // Author: Kristian Lytje
 
 #include <mini/detail/Landscape.h>
+
 #include <mini/detail/Evaluation.h>
 #include <utility/Exceptions.h>
 
@@ -19,11 +20,11 @@ SimpleDataset mini::Landscape::as_dataset() const {
     if (evals.empty()) {throw except::bad_order("Landscape::as_dataset: Cannot get evaluated points before a minimization call has been made.");}
     if (evals.front().vals.size() != 1) {throw except::invalid_operation("Landscape::as_dataset: Only 1D landscapes are convertible to datasets.");}
 
-    unsigned int N = evals.size();
+    int N = static_cast<int>(evals.size());
     std::vector<double> x(N), y(N);
     bool ordered = true;
 
-    for (unsigned int i = 0; i < N; i++) {
+    for (int i = 0; i < N; i++) {
         x[i] = evals[i].vals[0];
         y[i] = evals[i].fval;
