+++ source/rigidbody/detail/SystemSpecification.cpp
@@ -20,10 +20,10 @@ SystemSpecification::SystemSpecification(observer_ptr<const Rigidbody> rigidbody
 
         // initialize symmetry parameters from the body's symmetries
         assert(
-            absolute_parameters.parameters[i].symmetry_pars.size() == initial_conformation[i].size_symmetry() 
+            static_cast<int>(absolute_parameters.parameters[i].symmetry_pars.size()) == initial_conformation[i].size_symmetry() 
             && "Symmetry parameter size mismatch with body symmetry size."
         );
-        for (int j = 0; j < static_cast<int>(initial_conformation[i].size_symmetry()); ++j) {
+        for (int j = 0; j < initial_conformation[i].size_symmetry(); ++j) {
             absolute_parameters.parameters[i].symmetry_pars[j] = initial_conformation[i].symmetry().get(j)->clone();
         }
     }
